+++ CHANGELOG.md
@@ -11,6 +11,10 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 - Browse OTU: Replace descendants endpoint for the same used on TaxonPages. Now this panel is available for all ranks
 
+### Fixed
+
+- 2 issues with taxon names autocomplete (internal and api/v1)
+
 ## [0.40.2] - 2024-04-09
 
 ### Added